requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==4.2.0
celery==5.3.5
cffi==1.16.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==42.0.8
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.2.7
django-author==1.2.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-import-export==3.0.2
django-import-export-celery==1.3
django-reset-migrations==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
gunicorn==20.1.0
kombu==5.3.3
Markdown==3.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
postgres==4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pycparser==2.22
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0.1
redis==4.4.2
setuptools==70.3.0
setuptools-git==1.2
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2024.1
vine==5.1.0
waitress==2.1.2
wcwidth==0.2.6
wheel==0.43.0
xlrd==2.0.1
xlwt==1.3.0
